feat: separate concerns
make header file to include in standard
move #defines to header file
